Rename: 예외가 터지면 종료되는 게 아니라 다시 입력받도록 구현
