+++ src/main/java/racingcar/utils/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
